+++ spi-samples-configcenter/spi-samples-configcenter-consul/src/main/java/org/apache/dubbo/spi/samples/configcenter/impl/DemoServiceImpl.java
@@ -17,10 +17,22 @@
  *
  */
 
-package org.apache.dubbo.spi.samples.registry.api;
+package org.apache.dubbo.spi.samples.configcenter.impl;
 
-public interface DemoService {
 
-    String sayHello(String name);
+import org.apache.dubbo.rpc.RpcContext;
+import org.apache.dubbo.spi.samples.configcenter.api.DemoService;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class DemoServiceImpl implements DemoService {
+
+    @Override
+    public String sayHello(String name) {
+        System.out.println("[" + new SimpleDateFormat("HH:mm:ss").format(new Date()) + "] Hello " + name +
+                ", request from consumer: " + RpcContext.getContext().getRemoteAddress());
+        return "Hello " + name + ", response from provider: " + RpcContext.getContext().getLocalAddress();
+    }
 
 }
