+++ spi-samples-configcenter/spi-samples-configcenter-consul/src/main/resources/spring/configcenter-provider.xml
@@ -31,8 +31,8 @@
 
     <dubbo:config-center address="consul://${consul.address:127.0.0.1}:${consul.port:8500}"/>
 
-    <bean id="demoService" class="org.apache.dubbo.spi.samples.registry.impl.DemoServiceImpl"/>
+    <bean id="demoService" class="org.apache.dubbo.spi.samples.configcenter.impl.DemoServiceImpl"/>
 
-    <dubbo:service interface="org.apache.dubbo.spi.samples.registry.api.DemoService" ref="demoService"/>
+    <dubbo:service interface="org.apache.dubbo.spi.samples.configcenter.api.DemoService" ref="demoService"/>
 
 </beans>
