+++ spi-samples-registry/spi-samples-registry-redis/case-configuration.yml
@@ -1,0 +1,58 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+timeout: 600
+
+services:
+  dubbo-spi-samples-registry-redis:
+    type: app
+    basedir: .
+    mainClass: org.apache.dubbo.spi.samples.registry.RedisProvider
+    systemProps:
+      - redis.address=redis
+      - redis.port=6379
+    waitPortsBeforeRun:
+      - redis:6379
+    waitTimeout: 100
+    checkPorts:
+      - 20880
+    checkLog: "dubbo service started"
+    depends_on:
+      - redis
+
+  dubbo-spi-samples-registry-redis-test:
+    type: test
+    basedir: .
+    tests:
+      - "**/*IT.class"
+    systemProps:
+      - dubbo.address=dubbo-spi-samples-registry-redis
+      - dubbo.port=20880
+      - redis.address=redis
+      - redis.port=2379
+    waitPortsBeforeRun:
+      - dubbo-spi-samples-registry-redis:2181
+      - dubbo-spi-samples-registry-redis:20880
+      - redis:6379
+    waitTimeout: 100
+    depends_on:
+      - redis
+      - dubbo-spi-samples-registry-redis
+
+  redis:
+    image: redis:6.2.4
+    expose:
+      - 6379
