+++ spi-samples-registry/spi-samples-registry-redis/src/main/java/org/apache/dubbo/spi/samples/redis/RedisConsumer.java
@@ -1,0 +1,38 @@
+/*
+ *
+ *   Licensed to the Apache Software Foundation (ASF) under one or more
+ *   contributor license agreements.  See the NOTICE file distributed with
+ *   this work for additional information regarding copyright ownership.
+ *   The ASF licenses this file to You under the Apache License, Version 2.0
+ *   (the "License"); you may not use this file except in compliance with
+ *   the License.  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package org.apache.dubbo.spi.samples.redis;
+
+
+import org.apache.dubbo.spi.samples.redis.api.DemoService;
+
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+public class RedisConsumer {
+
+    public static void main(String[] args) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/configcenter-consumer.xml"});
+        context.start();
+        DemoService demoService = context.getBean("demoService", DemoService.class);
+
+        String hello = demoService.sayHello("world");
+        System.out.println(hello);
+
+    }
+}
