+++ spi-samples-registry/spi-samples-registry-redis/src/main/resources/spring/configcenter-provider.xml
@@ -25,12 +25,14 @@
        http://dubbo.apache.org/schema/dubbo http://dubbo.apache.org/schema/dubbo/dubbo.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
     <context:property-placeholder/>
 
-    <dubbo:application name="dubbo-spi-configcenter-etcd-consumer"/>
+    <dubbo:application name="dubbo-spi-registry-redis-provider"/>
 
-    <dubbo:registry address="consul://${consul.address:127.0.0.1}:${consul.port:8500}"/>
+    <dubbo:registry address="redis://${redis.address:127.0.0.1}:${redis.port:6379}"/>
 
-    <dubbo:config-center highest-priority="false" address="consul://${consul.address:127.0.0.1}:${consul.port:8500}"/>
+    <dubbo:config-center address="redis://${redis.address:localhost}:${redis.port:6379}"/>
 
-    <dubbo:reference id="demoService" interface="org.apache.dubbo.spi.samples.configcenter.api.DemoService"/>
+    <bean id="demoService" class="org.apache.dubbo.spi.samples.redis.impl.DemoServiceImpl"/>
+
+    <dubbo:service interface="org.apache.dubbo.spi.samples.redis.api.DemoService" ref="demoService"/>
 
 </beans>
